backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
anthropic>=0.40.0
python-dotenv==1.0.1
websockets==12.0
gtts==2.5.1
